+++ zwatershed/c_frontend.h
@@ -24,7 +24,7 @@ struct ZwatershedState {
 };
 
 std::vector<Metrics> process_thresholds(
-		const std::vector<size_t>& thresholds,
+		const std::vector<float>& thresholds,
 		size_t width, size_t height, size_t depth,
 		const float* affinity_data,
 		const std::vector<uint64_t*>& segmentation_data,
